event-consumers/employee-tracker/requirements.txt
kafka-python==1.4.7
python-dotenv==0.19.2
pydantic==1.10.13
mysql-connector-python==8.0.33
pytest==7.4.3
pytest-cov==4.1.0
redis==5.0.1 